whisper-local/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
python-multipart==0.0.6
aiofiles==23.2.0
requests==2.31.0
